requirements.txt
django~=2.0

path.py==2.4.1
ipdb==0.10.2
ipython==5.3.0
gunicorn==19.3.0
dj-database-url==0.4.1
psycopg2

# REST
django-rest-framework-mongoengine
djangorestframework
django-filter

django-cors-headers

whitenoise==3.3

Pillow

# storage S3
# django-storages
boto3
django-storages
# git+https://github.com/jschneier/django-storages.git#egg=django-storages

celery[redis]

# tests
coveralls
flake8
